Switch to gypsum backend for saving and loading datasets. (#42)
Similarly, we switch to the alabaster framework to store the data, rather than
RDS files. This should allow the files to be re-used in other languages.

Users can now grab SCEs directly from the gypsum backend with fetchDataset, and
they can perform listing and advanced searches directly on gypsum's SQLite
files. This lays the groundwork for removing dependencies on ExperimentHub.

All of the little per-dataset functions now call fetchDataset to retrieve
content, where it is back-compatible to do so. We plan to soft-deprecate all
per-dataset functions to simplify maintenance; users should just call
fetchDataset instead, and be responsible for any additional transformations.

We also provide functions for users to save and upload their own datasets.
This should simplify the process of making a contribution.